requirements.txt
streamlit
yfinance
pandas
requests
beautifulsoup4
google-genai
groq
plotly
toml
